+++ src/plugins/dev-mode/check-query.ts
@@ -1,13 +1,15 @@
 import type {
     RxPluginPreCreateRxQueryArgs,
     MangoQuery,
-    RxPluginPrePrepareQueryArgs
+    RxPluginPrePrepareQueryArgs,
+    DexiePreparedQuery
 } from '../../types';
 import deepEqual from 'fast-deep-equal';
 import { newRxError, newRxTypeError } from '../../rx-error';
 import {
     massageSelector
 } from 'pouchdb-selector-core';
+import { RxStorageDexieStatics } from '../dexie';
 
 /**
  * accidentially passing a non-valid object into the query params
@@ -44,6 +46,23 @@ export function checkQuery(args: RxPluginPreCreateRxQueryArgs) {
             });
         }
     });
+
+    // do not allow skip or limit for count queries
+    if (
+        args.op === 'count' &&
+        (
+            args.queryObj.limit ||
+            args.queryObj.skip
+        )
+    ) {
+        throw newRxError(
+            'QU15',
+            {
+                collection: args.collection.name,
+                query: args.queryObj
+            }
+        );
+    }
 }
 
 
@@ -89,6 +108,27 @@ export function checkMangoQuery(args: RxPluginPrePrepareQueryArgs) {
                     schema
                 }
             );
+        }
+    }
+
+
+    /**
+     * Ensure that a count() query can only be used
+     * with selectors that are fully satisfied by the used index.
+     */
+    if (args.rxQuery.op === 'count') {
+        const preparedQuery: DexiePreparedQuery<any> = RxStorageDexieStatics.prepareQuery(
+            args.rxQuery.collection.schema.jsonSchema,
+            args.mangoQuery
+        );
+        if (
+            !preparedQuery.queryPlan.selectorSatisfiedByIndex &&
+            !args.rxQuery.collection.database.allowSlowCount
+        ) {
+            throw newRxError('QU14', {
+                collection: args.rxQuery.collection,
+                query: args.mangoQuery
+            });
         }
     }
 }